requirements-dev.txt
# Development & testing dependencies
Faker>=22.0.0
django-extensions
# Add other development dependencies here like:
# pytest
# black
# flake8
# django-debug-toolbar 
